third_party/blink/web_tests/fast/dom/prototypes-expected.txt

This test checks some DOM object prototypes that WebKit has gotten wrong in the past. If the test passes, you'll see a series of 'PASS' messages below.

PASS: Object.prototype.isPrototypeOf(window.DOMException) should be true and is.
PASS: Object.prototype.isPrototypeOf(window.Option) should be true and is.
PASS: Object.prototype.isPrototypeOf(styleSheetList) should be true and is.
PASS: Object.prototype.isPrototypeOf(styleSheet) should be true and is.
PASS: Object.prototype.isPrototypeOf(cssRule) should be true and is.
PASS: Object.prototype.isPrototypeOf(nodeList) should be true and is.
PASS: Object.prototype.isPrototypeOf(namedNodesCollection) should be true and is.
